refactor(runner): move schemaToTypeString tests to dedicated file
Extract schemaToTypeString tests from llm-dialog-helpers.test.ts into
schema-format.test.ts for better organization.
